backend/test-requirements.txt
requests
pytest
pytest-subtests
python-dotenv
faker
